兴趣点航线定时删除加入判断逻辑；巡视结果返回分析图url,显示分析图

+++ cloud_api_serve_git/sample/src/main/java/com/dji/sample/df/wind/timer/DeleteWaylineFileTimer.java
@@ -63,7 +63,7 @@ public class DeleteWaylineFileTimer {
                 fileName = fileName.substring(0, i);
                 WaylineFileEntity waylineFileEntity = waylineFileMapper.selectOne(new LambdaQueryWrapper<WaylineFileEntity>().
                         eq(WaylineFileEntity::getName, fileName));
-                if (waylineFileEntity != null && !fileName.contains("top")) {
+                if (waylineFileEntity != null && !fileName.contains("top")&& !fileName.contains("poi")) {
 //                  删除数据库
                     waylineFileMapper.deleteById(waylineFileEntity.getId());
 //                  删除minio
